requirements.txt
# --- LangChain Stack Moderno ---
langchain>=0.2.0
langgraph>=0.1.0
langchain-ollama>=0.1.0
pydantic>=2.0.0

# --- Visualización ---
matplotlib

# --- Otras dependencias ---
pandas
python-dotenv
psycopg2-binary